feat(ui-kit): ship per-component entries so consumers can tree-shake
A consumer importing one component got the whole kit. Vite library mode
replaces tsup: one build entry per component from a new public.ts in each
component directory, plus the barrel, with each chunk importing its own
stylesheet via vite-plugin-lib-inject-css. The barrel stays the
documented import and still tree-shakes — dist/index.js is 3.66 kB of
re-exports over 19 separate chunk files rather than one bundle.

Measured on a Button-only consumer page with React externalized, built
against the packed tarball. The result depended entirely on the
consumer's bundler, which is why one measurement was not enough:

  raw esbuild   JS 233,672 B -> 9,383 B
  webpack 5     JS  62,924 B -> 8,581 B
  Vite/Rollup   already eliminated it; unchanged
  CSS (Vite)    33.06 kB -> 4.38 kB, 5.17 -> 1.08 kB gzipped

What the old build actually retained was Base UI's own submodules —
SelectRoot, MenuRoot, ToastRoot, FloatingFocusManager, @floating-ui/dom,
roughly 290 files — none of them reachable from Button, all of them
imported unconditionally somewhere in the single dist/index.js. Rollup
cascades "this wrapper is unused" through to "so is its exclusive
import"; esbuild's bundler does not. The /* @__PURE__ */ annotation on
the cva() calls that this was first attributed to fixes none of it:
byte-identical output with and without it. CSS was broken everywhere
regardless, having no per-component granularity to prune against.

Acceptance grows from one leg to four, because a Vite-only gate proves
the CSS half and silently skips the JS half:

  [vite]             the ecosystem's own tooling
  [esbuild-barrel]   JS absence, where the 25x regression lived
  [esbuild-subpath]  CSS absence — esbuild does not tree-shake CSS from a
                     barrel import at all, so this direction is only
                     assertable via a subpath
  [types]            tsc under both nodenext and bundler

Each assertion was checked by breaking it: the two that matter most were
proven against the real pre-repackaging tarball rather than a synthetic
mutation, so the gate is known to catch the regression it exists for.

Public contract changes: ./styles.css is gone, which is the point — CSS
now arrives with the component. CJS is dropped, ESM only, as the sibling
react-kit does. New ./* subpath exports, with ./components/* and
./__test-utils__/* explicitly denied so a deep path that type-checks
cannot resolve to a file that was never built.

Not adopted from react-kit: publishConfig.directory with a generated
package.json. Exports point straight into dist/ from the real package
root so the existing publish workflow keeps working.

The build tooling now type-checks too, under its own tsconfig — the main
project is scoped to src/ because vite-plugin-dts reads it to decide what
declarations to emit. The build is more code than tsup's single config
object now, and it is the code that decides what ships.
