+++ packages/ui-kit/package.json
@@ -3,17 +3,19 @@
   "version": "0.3.0-alpha.0",
   "description": "Standard React component base for FrontX templates - Base UI behavior, CSS Modules styling on semantic tokens, CVA variants",
   "type": "module",
-  "main": "./dist/index.cjs",
-  "module": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "exports": {
     ".": {
       "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "require": "./dist/index.cjs"
+      "import": "./dist/index.js"
     },
-    "./theme.css": "./dist/theme.css",
-    "./styles.css": "./dist/index.css"
+    "./components/*": null,
+    "./__test-utils__/*": null,
+    "./*": {
+      "types": "./dist/*.d.ts",
+      "import": "./dist/*.js"
+    },
+    "./theme.css": "./dist/theme.css"
   },
   "files": [
     "dist",
@@ -23,8 +25,9 @@
     "NOTICE"
   ],
   "scripts": {
-    "build": "tsup",
-    "type-check": "tsc --noEmit",
+    "build": "vite build",
+    "type-check": "tsc --noEmit && npm run type-check:tools",
+    "type-check:tools": "tsc -p tsconfig.tools.json",
     "type-check:test": "tsc -p tsconfig.test.json",
     "test": "npm run test:unit",
     "test:unit": "vitest --run --passWithNoTests=false",
@@ -44,16 +47,21 @@
     "@types/node": "^25.5.0",
     "@types/react": "19.2.17",
     "@types/react-dom": "19.2.3",
+    "@vitejs/plugin-react": "4.3.4",
     "@vitest/coverage-v8": "4.1.4",
+    "esbuild": "0.25.12",
     "jsdom": "26.1.0",
     "react": "19.2.8",
     "react-dom": "19.2.8",
-    "tsup": "^8.0.0",
+    "rollup-plugin-node-externals": "^9.0.1",
     "typescript": "^5.4.2",
+    "vite": "^6.4.3",
+    "vite-plugin-dts": "^4.5.4",
+    "vite-plugin-lib-inject-css": "^2.2.2",
     "vitest": "4.1.4"
   },
   "sideEffects": [
-    "*.css"
+    "**/*.css"
   ],
   "private": true,
   "publishConfig": {
