+++ packages/ui-kit/src/components/badge/public.ts
@@ -1,0 +1,1 @@
+export { Badge, type BadgeProps } from './badge.js';
