+++ packages/ui-kit/src/components/button/public.ts
@@ -1,0 +1,1 @@
+export { Button, type ButtonProps } from './button.js';
