+++ packages/ui-kit/src/components/card/public.ts
@@ -1,0 +1,16 @@
+export {
+  Card,
+  type CardProps,
+  CardAction,
+  type CardActionProps,
+  CardContent,
+  type CardContentProps,
+  CardDescription,
+  type CardDescriptionProps,
+  CardFooter,
+  type CardFooterProps,
+  CardHeader,
+  type CardHeaderProps,
+  CardTitle,
+  type CardTitleProps,
+} from './card.js';
