+++ packages/ui-kit/src/components/checkbox/public.ts
@@ -1,0 +1,1 @@
+export { Checkbox, type CheckboxProps } from './checkbox.js';
