+++ packages/ui-kit/src/components/dropdown-menu/public.ts
@@ -1,0 +1,28 @@
+export {
+  DropdownMenu,
+  DropdownMenuCheckboxItem,
+  type DropdownMenuCheckboxItemProps,
+  DropdownMenuContent,
+  type DropdownMenuContentProps,
+  DropdownMenuGroup,
+  type DropdownMenuGroupProps,
+  DropdownMenuItem,
+  type DropdownMenuItemProps,
+  DropdownMenuLabel,
+  type DropdownMenuLabelProps,
+  DropdownMenuRadioGroup,
+  type DropdownMenuRadioGroupProps,
+  DropdownMenuRadioItem,
+  type DropdownMenuRadioItemProps,
+  DropdownMenuSeparator,
+  type DropdownMenuSeparatorProps,
+  DropdownMenuShortcut,
+  type DropdownMenuShortcutProps,
+  DropdownMenuSub,
+  DropdownMenuSubContent,
+  type DropdownMenuSubContentProps,
+  DropdownMenuSubTrigger,
+  type DropdownMenuSubTriggerProps,
+  DropdownMenuTrigger,
+  type DropdownMenuTriggerProps,
+} from './dropdown-menu.js';
