+++ packages/ui-kit/src/components/field/public.ts
@@ -1,0 +1,10 @@
+export {
+  Field,
+  type FieldProps,
+  FieldDescription,
+  type FieldDescriptionProps,
+  FieldError,
+  type FieldErrorProps,
+  FieldLabel,
+  type FieldLabelProps,
+} from './field.js';
