+++ packages/ui-kit/src/components/input/public.ts
@@ -1,0 +1,1 @@
+export { Input, type InputProps } from './input.js';
