+++ packages/ui-kit/src/components/label/public.ts
@@ -1,0 +1,1 @@
+export { Label, type LabelProps } from './label.js';
