+++ packages/ui-kit/src/components/radio-group/public.ts
@@ -1,0 +1,6 @@
+export {
+  RadioGroup,
+  type RadioGroupProps,
+  RadioGroupItem,
+  type RadioGroupItemProps,
+} from './radio-group.js';
