+++ packages/ui-kit/src/components/select/public.ts
@@ -1,0 +1,19 @@
+export {
+  Select,
+  SelectContent,
+  type SelectContentProps,
+  SelectGroup,
+  type SelectGroupProps,
+  SelectItem,
+  type SelectItemProps,
+  SelectLabel,
+  type SelectLabelProps,
+  SelectScrollDownButton,
+  SelectScrollUpButton,
+  SelectSeparator,
+  type SelectSeparatorProps,
+  SelectTrigger,
+  type SelectTriggerProps,
+  SelectValue,
+  type SelectValueProps,
+} from './select.js';
