+++ packages/ui-kit/src/components/separator/public.ts
@@ -1,0 +1,1 @@
+export { Separator, type SeparatorProps } from './separator.js';
