+++ packages/ui-kit/src/components/skeleton/public.ts
@@ -1,0 +1,1 @@
+export { Skeleton, type SkeletonProps } from './skeleton.js';
