+++ packages/ui-kit/src/components/switch/public.ts
@@ -1,0 +1,1 @@
+export { Switch, type SwitchProps } from './switch.js';
