+++ packages/ui-kit/src/components/tabs/public.ts
@@ -1,0 +1,10 @@
+export {
+  Tabs,
+  type TabsProps,
+  TabsContent,
+  type TabsContentProps,
+  TabsList,
+  type TabsListProps,
+  TabsTrigger,
+  type TabsTriggerProps,
+} from './tabs.js';
