+++ packages/ui-kit/src/components/textarea/public.ts
@@ -1,0 +1,1 @@
+export { Textarea, type TextareaProps } from './textarea.js';
