+++ packages/ui-kit/src/components/toast/public.ts
@@ -1,0 +1,7 @@
+export {
+  createToastManager,
+  toast,
+  Toaster,
+  type ToasterProps,
+  useToastManager,
+} from './toast.js';
