+++ packages/ui-kit/src/components/tooltip/public.ts
@@ -1,0 +1,9 @@
+export {
+  Tooltip,
+  TooltipContent,
+  type TooltipContentProps,
+  TooltipProvider,
+  type TooltipProviderProps,
+  TooltipTrigger,
+  type TooltipTriggerProps,
+} from './tooltip.js';
