+++ packages/ui-kit/src/index.ts
@@ -3,155 +3,45 @@
  * templates.
  *
  * Styles ship separately: import '@gears-frontx/ui-kit/theme.css' (design
- * tokens) and '@gears-frontx/ui-kit/styles.css' (compiled component styles)
- * once in the consumer entry module.
+ * tokens) once in the consumer entry module. Component CSS ships alongside
+ * each component's own chunk (see vite.config.ts / scripts/buildPlugin.ts)
+ * and needs no separate import — importing a component pulls its styles in
+ * with it.
+ *
+ * This barrel re-exports every component's public.ts rather than declaring
+ * exports itself. Each public.ts is also its own build entry (globbed in
+ * scripts/buildPlugin.ts), so `export * from './components/x/public'` here
+ * re-exports an already-separate chunk instead of inlining one — that's what
+ * keeps this barrel tree-shakeable: a consumer importing only `Button` gets
+ * a thin dist/index.js that re-exports dist/button.js and 18 sibling chunks
+ * it never touches, and their bundler can drop the ones it never imports.
+ *
+ * Relative specifiers below carry an explicit `.js` extension even though
+ * the source files are `.ts` — required because this flows straight into
+ * the emitted `dist/index.d.ts`, and a consumer building under
+ * `moduleResolution: "nodenext"` rejects extensionless relative specifiers
+ * in an ESM package's declarations (TS2305, every symbol invisible). The
+ * `.js` specifier still resolves to the local `.ts` file under our own
+ * `moduleResolution: "bundler"` (and under Vite/vitest, which special-case
+ * a `.js` specifier resolving to a sibling `.ts`/`.tsx`) — it is not a
+ * claim that a `.js` file exists here.
  */
-export { Badge, type BadgeProps } from './components/badge/badge';
-export { Button, type ButtonProps } from './components/button/button';
-export {
-  Card,
-  type CardProps,
-  CardAction,
-  type CardActionProps,
-  CardContent,
-  type CardContentProps,
-  CardDescription,
-  type CardDescriptionProps,
-  CardFooter,
-  type CardFooterProps,
-  CardHeader,
-  type CardHeaderProps,
-  CardTitle,
-  type CardTitleProps,
-} from './components/card/card';
-export { Checkbox, type CheckboxProps } from './components/checkbox/checkbox';
-export {
-  Dialog,
-  DialogClose,
-  type DialogCloseProps,
-  DialogContent,
-  type DialogContentProps,
-  DialogDescription,
-  type DialogDescriptionProps,
-  DialogFooter,
-  type DialogFooterProps,
-  DialogHeader,
-  type DialogHeaderProps,
-  DialogTitle,
-  type DialogTitleProps,
-  DialogTrigger,
-  type DialogTriggerProps,
-} from './components/dialog/dialog';
-export {
-  DropdownMenu,
-  DropdownMenuCheckboxItem,
-  type DropdownMenuCheckboxItemProps,
-  DropdownMenuContent,
-  type DropdownMenuContentProps,
-  DropdownMenuGroup,
-  type DropdownMenuGroupProps,
-  DropdownMenuItem,
-  type DropdownMenuItemProps,
-  DropdownMenuLabel,
-  type DropdownMenuLabelProps,
-  DropdownMenuRadioGroup,
-  type DropdownMenuRadioGroupProps,
-  DropdownMenuRadioItem,
-  type DropdownMenuRadioItemProps,
-  DropdownMenuSeparator,
-  type DropdownMenuSeparatorProps,
-  DropdownMenuShortcut,
-  type DropdownMenuShortcutProps,
-  DropdownMenuSub,
-  DropdownMenuSubContent,
-  type DropdownMenuSubContentProps,
-  DropdownMenuSubTrigger,
-  type DropdownMenuSubTriggerProps,
-  DropdownMenuTrigger,
-  type DropdownMenuTriggerProps,
-} from './components/dropdown-menu/dropdown-menu';
-export {
-  Field,
-  type FieldProps,
-  FieldDescription,
-  type FieldDescriptionProps,
-  FieldError,
-  type FieldErrorProps,
-  FieldLabel,
-  type FieldLabelProps,
-} from './components/field/field';
-export { Input, type InputProps } from './components/input/input';
-export { Label, type LabelProps } from './components/label/label';
-export {
-  RadioGroup,
-  type RadioGroupProps,
-  RadioGroupItem,
-  type RadioGroupItemProps,
-} from './components/radio-group/radio-group';
-export {
-  Select,
-  SelectContent,
-  type SelectContentProps,
-  SelectGroup,
-  type SelectGroupProps,
-  SelectItem,
-  type SelectItemProps,
-  SelectLabel,
-  type SelectLabelProps,
-  SelectScrollDownButton,
-  SelectScrollUpButton,
-  SelectSeparator,
-  type SelectSeparatorProps,
-  SelectTrigger,
-  type SelectTriggerProps,
-  SelectValue,
-  type SelectValueProps,
-} from './components/select/select';
-export { Separator, type SeparatorProps } from './components/separator/separator';
-export { Skeleton, type SkeletonProps } from './components/skeleton/skeleton';
-export { Switch, type SwitchProps } from './components/switch/switch';
-export {
-  Table,
-  type TableProps,
-  TableBody,
-  type TableBodyProps,
-  TableCaption,
-  type TableCaptionProps,
-  TableCell,
-  type TableCellProps,
-  TableFooter,
-  type TableFooterProps,
-  TableHead,
-  type TableHeadProps,
-  TableHeader,
-  type TableHeaderProps,
-  TableRow,
-  type TableRowProps,
-} from './components/table/table';
-export {
-  Tabs,
-  type TabsProps,
-  TabsContent,
-  type TabsContentProps,
-  TabsList,
-  type TabsListProps,
-  TabsTrigger,
-  type TabsTriggerProps,
-} from './components/tabs/tabs';
-export { Textarea, type TextareaProps } from './components/textarea/textarea';
-export {
-  createToastManager,
-  toast,
-  Toaster,
-  type ToasterProps,
-  useToastManager,
-} from './components/toast/toast';
-export {
-  Tooltip,
-  TooltipContent,
-  type TooltipContentProps,
-  TooltipProvider,
-  type TooltipProviderProps,
-  TooltipTrigger,
-  type TooltipTriggerProps,
-} from './components/tooltip/tooltip';
+export * from './components/badge/public.js';
+export * from './components/button/public.js';
+export * from './components/card/public.js';
+export * from './components/checkbox/public.js';
+export * from './components/dialog/public.js';
+export * from './components/dropdown-menu/public.js';
+export * from './components/field/public.js';
+export * from './components/input/public.js';
+export * from './components/label/public.js';
+export * from './components/radio-group/public.js';
+export * from './components/select/public.js';
+export * from './components/separator/public.js';
+export * from './components/skeleton/public.js';
+export * from './components/switch/public.js';
+export * from './components/table/public.js';
+export * from './components/tabs/public.js';
+export * from './components/textarea/public.js';
+export * from './components/toast/public.js';
+export * from './components/tooltip/public.js';
